requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
email-validator
multipart